requirements.txt
Flask==2.0.1
beautifulsoup4==4.9.3
requests==2.25.1
pymongo==3.11.4
lxml==4.6.3
Werkzeug==2.0.1
python-dateutil==2.8.1
bleach==3.1.5
Celery==5.2.3
redis==4.1.0
Flask-Cors==3.0.10
